检查是否是本地qemu问题

+++ userprog/process.c
@@ -14,7 +14,7 @@ extern void user_ret(void);
 
 void create_user_vaddr_bitmap(struct task_struct* user_prog) {
    user_prog->usrprog_vaddr.vaddr_start = USER_VADDR_START;
-   uint32_t bitmap_pg_cnt = DIV_ROUND_UP((USER_STACK - USER_VADDR_START) / PAGESIZE / 8 , PAGESIZE);
+   uint64_t bitmap_pg_cnt = DIV_ROUND_UP((USER_STACK - USER_VADDR_START) / PAGESIZE / 8 , PAGESIZE);
    user_prog->usrprog_vaddr.btmp.bits = (uint8_t*)get_pages(bitmap_pg_cnt);
    user_prog->usrprog_vaddr.btmp.btmp_bytes_len = (USER_STACK - USER_VADDR_START) / PAGESIZE / 8;
    bitmap_init(&user_prog->usrprog_vaddr.btmp);
@@ -66,5 +66,6 @@ void page_dir_activate(struct task_struct* pcb)
 		write_csr_pgdl(pcb->pgdir);
 		invalidate();
 	}
+	//printk("next pgdir=%x\n",pcb->pgdir);
 }
